update test fixes
