+++ tests/Oryx.Integration.Tests/Node/NodeOtherEndtoEndTests.cs
@@ -247,47 +247,6 @@ namespace Microsoft.Oryx.Integration.Tests
                 });
         }
 
-        [Theory]
-        [MemberData(nameof(TestValueGenerator.GetZipOptions_NodeVersions), MemberType = typeof(TestValueGenerator))]
-        public async Task CanBuildAndRun_NodeApp(string nodeVersion)
-        {
-            // Arrange
-            var appName = "webfrontend";
-            var volume = CreateAppVolume(appName);
-            var appDir = volume.ContainerDir;
-            var buildScript = new ShellScriptBuilder()
-               .AddCommand($"oryx build {appDir} --platform nodejs --language-version {nodeVersion}")
-               .ToString();
-            var runScript = new ShellScriptBuilder()
-                .AddCommand($"oryx -appPath {appDir} -bindPort {ContainerPort}")
-                .AddCommand(DefaultStartupFilePath)
-                .ToString();
-
-            await EndToEndTestHelper.BuildRunAndAssertAppAsync(
-                appName,
-                _output,
-                volume,
-                "/bin/sh",
-                new[]
-                {
-                    "-c",
-                    buildScript
-                },
-                $"oryxdevmcr.azurecr.io/public/oryx/node-{nodeVersion}",
-                ContainerPort,
-                "/bin/sh",
-                new[]
-                {
-                    "-c",
-                    runScript
-                },
-                async (hostPort) =>
-                {
-                    var data = await _httpClient.GetStringAsync($"http://localhost:{hostPort}/");
-                    Assert.Contains("Say It Again", data);
-                });
-        }
-
         [Fact]
         public async Task NodeStartupScript_UsesPortEnvironmentVariableValue()
         {
